Divide the options window in 2 tabs.
- A colors tab, where you can change the ruler elements colors.
- A general tab, with the rest of the controls.
- The options window now is only initiated once on load.
